+++ packages/spotlight/src/spotlight.js
@@ -203,7 +203,7 @@ const Spotlight = (function () {
 			if (currentFocusedElement) {
 				currentFocusedElement.blur();
 			}
-			elem.focus();
+			elem.focus({preventScroll: true});
 			focusChanged(elem, containerIds);
 		};
 
@@ -224,7 +224,7 @@ const Spotlight = (function () {
 			currentFocusedElement.blur();
 		}
 
-		elem.focus();
+		elem.focus({preventScroll: true});
 
 		_duringFocusChange = false;