+++ packages/moonstone/ExpandableItem/ExpandableTransitionContainer.js
@@ -11,6 +11,7 @@ import Transition from '@enact/ui/Transition';
  */
 
 const ExpandableTransitionContainerBase = SpotlightContainerDecorator(
+	{preventScroll: true},
 	Resizable(
 		{resize: 'onTransitionEnd', filter: (ev) => ev.propertyName === 'height'},
 		Transition