+++ packages/moonstone/Scrollable/ScrollableNative.js
@@ -903,6 +903,7 @@ const ScrollableNative = Skinnable(SpotlightContainerDecorator(
 	{
 		overflow: true,
 		preserveId: true,
+		preventScroll: true,
 		restrict: 'self-first'
 	},
 	ScrollableBaseNative