+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` to prevent the browser from scrolling elements into view when focused within a spotlight container configured with `overflow: true`
+
 ## [2.1.1] - 2018-08-27
 
 ### Fixed