+++ packages/spotlight/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ## [2.1.1] - 2018-08-27
 
+### Changed
+
+- `spotlight` to call focus() function of the target element with `{preventScroll: true}` option
+
 ### Fixed
 
 - `spotlight` to correctly handle focus with `'self-only'` containers